requirements.txt

# SOaC Framework Dependencies
# Python 3.11+

# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# API and web frameworks
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
jsonschema>=4.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.7.0
flake8>=6.1.0
pylint>=2.17.0

# Security
cryptography>=41.0.0

# Optional: For advanced features
# scikit-learn>=1.3.0  # For ML-based threat detection
# redis>=4.6.0  # For caching
# celery>=5.3.0  # For async task processing
